refactor(Tab): remove the interface to index file
