--- src/components/navigation/Tab/interfaces/Tab.interface.ts
@@ -1,7 +1,0 @@
-export interface ITabProps {
-  label: string;
-  id: string;
-  isDisabled: boolean;
-  isSelected: boolean;
-  handleClick: () => void;
-}
